# Conflicts:
#	WebApp_Doctor/obj/Debug/net6.0/WebApp_Doctor.csproj.AssemblyReference.cache

+++ Hart_Check_Official/Models/ArchievedRecord.cs
@@ -11,8 +11,8 @@ namespace Hart_Check_Official.Models
         public string lastName { get; set; }
         public string email { get; set; }
         public DateTime birthdate { get; set; }
-        public int systolic { get; set; }
-        public int diastolic { get; set; }
+        public double systolic { get; set; }
+        public double diastolic { get; set; }
 
         public Patients patients { get; set; }
     }
